suggestion(server): organize files in server directory
