+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './client'
 export * from './server';
-export * from './shared/types'
