+++ src/server/utils/get-query-param-value.ts
@@ -1,0 +1,12 @@
+import { UniversalRequest } from '../types';
+
+/**
+ * Возвращает значение query-параметра по ключу.
+ */
+export function getQueryParamValue(request: UniversalRequest, queryKey: string) {
+    if (!request.url) {
+        return null;
+    }
+
+    return new URL(request.url).searchParams.get(queryKey);
+}
